requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2              # Async MongoDB driver
pymongo==4.6.0            # MongoDB driver

# Data Models
pydantic==2.5.0
pydantic-settings==2.1.0

# File Processing
PyPDF2==3.0.1             # PDF text extraction
python-docx==1.1.0        # DOCX text extraction
python-magic==0.4.27      # File type detection

# Security & Validation
python-multipart==0.0.6   # File upload support

# Utilities
python-dotenv==1.0.0      # Environment variables
aiofiles==23.2.1          # Async file operations

# Logging & Monitoring
structlog==23.2.0         # Structured logging

# Development Dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2             # For testing HTTP endpoints
pytest-cov==4.1.0         # Coverage reporting

# Optional AI Dependencies (for future use)
# openai==1.3.0
# anthropic==0.8.0